+++ web/modules/custom/hoeringsportal_project/modules/hoeringsportal_project_fixtures/src/Fixture/ProjectLandingPageFixture.php
@@ -1,0 +1,72 @@
+<?php
+
+namespace Drupal\hoeringsportal_project_fixtures\Fixture;
+
+use Drupal\content_fixtures\Fixture\AbstractFixture;
+use Drupal\content_fixtures\Fixture\DependentFixtureInterface;
+use Drupal\content_fixtures\Fixture\FixtureGroupInterface;
+use Drupal\hoeringsportal_base_fixtures\Fixture\MediaFixture;
+use Drupal\node\Entity\Node;
+use Drupal\paragraphs\Entity\Paragraph;
+
+/**
+ * Landing page fixture.
+ *
+ * @package Drupal\hoeringsportal_hearing_fixtures\Fixture
+ */
+class ProjectLandingPageFixture extends AbstractFixture implements DependentFixtureInterface, FixtureGroupInterface {
+
+  /**
+   * {@inheritdoc}
+   */
+  public function load() {
+    $entity = Node::create([
+      'type' => 'landing_page',
+      'title' => 'Projekter',
+      'field_teaser' => [
+        'value' => <<<'BODY'
+<p>Samler information om projekter i kommunen.</p>
+BODY
+      ],
+      'field_media_image_single' => [$this->getReference('media:Large3')],
+    ]);
+
+    $paragraph = Paragraph::create([
+      'type' => 'content_list',
+      'field_content_list' => [
+        'target_id' => 'all_projects',
+        'display_id' => 'default',
+        'data' => '',
+      ],
+      'field_list_title' => 'All projects',
+    ]);
+    $paragraph->save();
+
+    $entity->set('field_section', [
+      [
+        'target_id' => $paragraph->id(),
+        'target_revision_id' => $paragraph->getRevisionId(),
+      ],
+    ]);
+
+    $entity->save();
+    $this->addReference('node:landing_page:Projects', $entity);
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getDependencies() {
+    return [
+      MediaFixture::class,
+    ];
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getGroups() {
+    return ['node', 'project'];
+  }
+
+}